feat(bot-architecture): separate EMA fetch logic + add BotIntentsCard and shared fetch utility
